Renamed new external-logging app to 'logger'.
