+++ getyour/logger/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class LogConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
-    name = "log_ext"
+    name = "logger"
